pyproject.toml
[tool.poetry]
name = "google-play-signer"
version = "1.0.0"
description = ""
authors = ["miguelchenriquess22 <miguel.henriques@aptoide.com>"]
readme = "README.md"
packages = [{include = "google_play_signer"}]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.0"
flask = "^2.3.2"
flask-restx = "^1.1.0"
google-api-python-client = "^2.86.0"
python-jose = "^3.3.0"
gunicorn = "^20.1.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
